% Auto-generated: do not edit by hand
\name{''CircuitMap}

\alias{''CircuitMap}

\title{CircuitMap component}

\description{
Assumes horizontal stack
}

\usage{
''CircuitMap(id=NULL, label=NULL, value=NULL, clickData=NULL,
hoverNodeData=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{value}{Named list. The value displayed in the input.}

\item{clickData}{Named list. Data from latest click event. Read-only.}

\item{hoverNodeData}{Named list. Data of hovered node}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

